% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zip.R
\name{download_zip}
\alias{download_zip}
\title{Downloads zipped file of modeling results}
\usage{
download_zip(project_id, path, filename)
}
\arguments{
\item{project_id}{id of the project to be downloaded which can be founded by using the function \code{list_files}..}

\item{path}{folder in which the downloaded files should be saved.}

\item{filename}{name of the zipped file.}
}
\value{
API response with information about saved file.
}
\description{
This function allows the user to download the \code{forecast_pack}'s for any model associated
to the \code{logged user}.
}
\examples{
\dontrun{
if(interactive()){

 ## EXAMPLE - Downloading project files from an example
 # Note that you will need to have a previously ran project
 project_id <- 'User project ID'
 folder_path <- 'Folder in which the file will be saved'
 filename <- 'Name of the file to be saved'

 faas4i::download_zip(project_id = project_id, path=folder_path, filename = filename)
 }
}
}
\seealso{
\code{\link[faas4i]{character(0)}}
 \code{\link[httr]{insensitive}},\code{\link[httr]{GET}},\code{\link[httr]{add_headers}},\code{\link[httr]{timeout}},\code{\link[httr]{write_disk}}
}
