% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_model_spec.R
\name{fill_model_spec}
\alias{fill_model_spec}
\title{Fill list with modeling specifications}
\usage{
fill_model_spec(model_spec)
}
\arguments{
\item{model_spec}{list with modeling and cross validation setup.}
}
\value{
\code{model_spec} complete with default settings for parameters not defined
}
\description{
This function fill empty parameters for \code{model_spec}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
