% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get Auth0 acess token}
\usage{
get_access_token(
  config_file = paste0(system.file(package = "faas4i"), "/config.json")
)
}
\arguments{
\item{config_file}{set the user email.}
}
\value{
auth0 token
}
\description{
Gets access token from the api that will be used for modeling and update.
}
\examples{
\dontrun{
if(interactive()){
 ## EXAMPLE - Downloading project files from an example

 my_token <- faas4i::get_access_token()
 }
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
