% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Getting url for modeling}
\usage{
get_url(type, body_size = NULL)
}
\arguments{
\item{type}{Modeling type, if 'models' or 'update'}

\item{body_size}{Size of compressed body, in megabits}
}
\value{
Api url.
}
\description{
Getting url for FaaS modeling.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
