% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{list_projects}
\alias{list_projects}
\title{Listing user's projects}
\usage{
list_projects(project_id = NULL, ...)
}
\arguments{
\item{project_id}{if provided, retrieves information for a specific project}

\item{...}{advanced parameters.}
}
\value{
List with information for all projects sent by the user.
}
\description{
Returns a list of projects that belong to the user, or the
information about a single project if \code{project_id} is provided.
}
\examples{
\dontrun{
if(interactive()){

 ## EXAMPLE - Getting list of projects
 # Note that you will need to have a previously ran project

 my_projects <- faas4i::list_projects()
 }
}
}
\seealso{
\code{\link[faas4i]{character(0)}}
 \code{\link[utils]{str}}
}
