% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{match_names}
\alias{match_names}
\title{Matching original names with forecast_pack formulations}
\usage{
match_names(dataset, names_fp)
}
\arguments{
\item{dataset}{dataset with names to be matched with \code{names_fp}.}

\item{names_fp}{names of variables to match in the \code{dataset}.}
}
\value{
The dataset with matched names.
}
\description{
This function will receive a dataset and a vector of all names
of variables from dataset that is used in the forecast_pack, and matches them
using any parameterization that we currently use at 4i (either front or api).
}
\details{
For variables not found in \code{names_fp}, we apply the api parameterization, to make sure we have clean names.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[base]{chartr}}
}
