% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_check.R
\name{package_version_check}
\alias{package_version_check}
\title{Checking package version for faas4i}
\usage{
package_version_check(
  repo = "https://github.com/4intelligence/faas4i",
  proxy_url,
  proxy_port
)
}
\arguments{
\item{repo}{repository url to obtain latest version}

\item{proxy_url}{The url to be used as proxy}

\item{proxy_port}{Proxy port number}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function checks if the package is running with the latest vesion available of faas4i.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[remotes]{install_github}}
}
