% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_body_update.R
\name{prepare_body_update}
\alias{prepare_body_update}
\title{Prepares body for model update}
\usage{
prepare_body_update(
  pack_list,
  date_variable,
  date_format,
  project_name,
  user_email,
  cv_update,
  model_spec,
  base_dates,
  force_request,
  outlier_update,
  breakdown
)
}
\arguments{
\item{pack_list}{list with information about all packs to be updated. For each pack a list with a \code{forecast_pack} and a \code{new_data} should be provided.}

\item{date_variable}{name of variable with date information in all \code{new_data} in \code{pack_list}.}

\item{date_format}{format of \code{date_variable} in all \code{new_data} in \code{pack_list}.}

\item{project_name}{project name. It accepts character and numeric inputs. Special characters will be removed.}

\item{user_email}{set the user email. We are going to use it to let you know when the modeling is over.}

\item{cv_update}{TRUE or FALSE, indicating whether cross validation should be ran again.}

\item{model_spec}{list containing: \code{fill_forecast}, \code{n_steps}, \code{n_windows} and \code{cv_summary}, see details for its description. All arguments are optional, however \code{n_steps}, \code{n_windows} and \code{cv_summary} should only be used when \code{cv_update = TRUE}.}

\item{base_dates}{TRUE or FALSE, indicating whether initial date in modeling should be kept for update, if possible.}

\item{force_request}{TRUE or FALSE, indicating whether estimation should continue when not all packs can be updated.}

\item{outlier_update}{TRUE or FALSE, indicating whether a new search for outliers should be conducted. Search for outliers will be done only when \code{cv_update} = TRUE.}

\item{breakdown}{TRUE, FALSE or vector with 'row_id' of models to run breakdown (max of 3 models). If \code{breakdown} is TRUE, model breakdown will be calculated for the first 3 arima models. If \code{breakdown} is FALSE, model breakdown from previous modeling will be removed from \code{forecast_pack}.}
}
\value{
List with encoded \code{forecast_pack} and \code{new_data}.
}
\description{
This function will prepare the encoded body to be sent via api after a series of consistency checks have been performed.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[jsonlite]{base64}},\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[base]{serialize}},\code{\link[base]{memCompress}}
 \code{\link[caTools]{base64encode & base64decode}}
}
