% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_R.R
\name{pretty_R}
\alias{pretty_R}
\title{Print messages from validation in a pretty format}
\usage{
pretty_R(message_list, extra_list = FALSE)
}
\arguments{
\item{message_list}{The list with messages, either error list or warning list}

\item{extra_list}{A boolean indicating if there is an extra layer of list (currently TRUE for error lists)}
}
\value{
A formmated string to be nicely printed
}
\description{
Ths function uses messages derived from valdation and organize the
input to be pretty printed
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
