% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_update.R
\name{run_update}
\alias{run_update}
\title{FaaS - updating models in forecast pack generated by scale modeling (discontinued)}
\usage{
run_update(...)
}
\arguments{
\item{...}{advanced parameters.}
}
\value{
A warning informing the user that this way of updating a model has been
discontinued
}
\description{
This way of updating a model has been discontinued. For details about
the new way to perform updates, see the documentation in
https://storage.googleapis.com/bkt-prod-4casthub/docs/model-update-api-en-us.pdf.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
