% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_treat.R
\name{string_treat}
\alias{string_treat}
\title{Treats string so it is a clean names}
\usage{
string_treat(string)
}
\arguments{
\item{string}{name to be cleaned.}
}
\value{
A clean name.
}
\description{
This function will receive a single string and apply treatments
to make it into a clean name. This replicates the process applied in our
front end.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
