% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_info_update.R
\name{validate_info_update}
\alias{validate_info_update}
\title{Validates info in forecast_pack and new_data for model_update}
\usage{
validate_info_update(
  index,
  forecast_pack,
  new_data,
  date_variable,
  date_format,
  base_dates
)
}
\arguments{
\item{index}{index of \code{forecast_pack} and \code{new_data} in \code{pack_list}.}

\item{forecast_pack}{pack outputted by FaaS with modeling information to be updated.}

\item{new_data}{dataset to be used to update models in \code{forecast_pack}. Variable names should be the same as inital modeling.}

\item{date_variable}{name of variable with date information in all \code{new_data} in \code{pack_list}.}

\item{date_format}{format of \code{date_variable} in all \code{new_data} in \code{pack_list}.}

\item{base_dates}{TRUE or FALSE, indicating whether initial date in modeling should be kept for update, if possible.}
}
\value{
A list with the following components:
\itemize{
   \item update_i - TRUE if \code{forecast_pack} and \code{new_data} passed all checks and can be updated, FALSE otherwise.
   \item new_data - same \code{new_data} inputted, but with clean names, as used in \code{forecast_pack}.
   \item txt - string with error messages, or empty string when no error is found.
}
}
\description{
This function will perform consistency checks for a single new_data
and forecast_pack, to guarantee that the user has provided appropriated parameters.
}
\details{
This function will perform the following checks:
\itemize{
   \item If \code{forecast_pack} has at least one non combination model, as we need to update models used to compose combination to be able to update combination.
   \item If \code{date_variable} and \code{date_format} were properly defined.
   \item If all variables used in original modeling in \code{forecast_pack} are available in \code{new_data}.
   \item If transformations applied in original modeling are still appropriate. Eg., if the variable was originally log transformed but assumes negative values in \code{new_data}, we are not able to update this pack.
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[base]{as.Date}},\code{\link[base]{funprog}},\code{\link[base]{merge}},\code{\link[base]{any}},\code{\link[base]{log}}
}
