% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_user_input.R
\name{validate_user_input}
\alias{validate_user_input}
\title{Validate User Input for FaaS modeling}
\usage{
validate_user_input(
  data_list,
  date_variable,
  date_format,
  model_spec,
  project_name,
  user_email,
  skip_validation
)
}
\arguments{
\item{data_list}{list with datasets to be modeled, where the list elements must be named after the dependent variable. You cannot have more than one dependent variable with same name in a \code{data_list}.}

\item{date_variable}{name of variable with date information in all datasets in \code{data_list}.}

\item{date_format}{format of \code{date_variable} in all datasets in \code{data_list}.}

\item{model_spec}{list with modeling and cross validation setup.}

\item{project_name}{project name. A string with character and/or numeric inputs that should be at most 50 characters long. Special characters will be removed.}

\item{user_email}{email to receive the outputs.}

\item{skip_validation}{TRUE or FALSE, indicating if validation should be skipped.}
}
\value{
None. Will break if any argument is not properly defined.
}
\description{
This function will perform an initial check if all necessary
arguments to run the \code{run_models} were provided.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
