% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_user_input_update.R
\name{validate_user_input_update}
\alias{validate_user_input_update}
\title{Validate User Input for Model Update}
\usage{
validate_user_input_update(
  pack_list,
  date_variable,
  date_format,
  project_name,
  user_email,
  cv_update,
  model_spec,
  base_dates,
  force_request,
  outlier_update,
  breakdown
)
}
\arguments{
\item{pack_list}{list with information about all packs to be updated. For each pack a list with a \code{forecast_pack} and a \code{new_data} should be provided.}

\item{date_variable}{name of variable with date information in all \code{new_data} in \code{pack_list}.}

\item{date_format}{format of \code{date_variable} in all \code{new_data} in \code{pack_list}.}

\item{project_name}{project name. A string with character and/or numeric inputs that should be at most 50 characters long. Special characters will be removed.}

\item{user_email}{email to receive the outputs.}

\item{cv_update}{TRUE or FALSE, indicating whether cross validation should be ran again.}

\item{model_spec}{list containing: \code{fill_forecast}, \code{n_steps}, \code{n_windows} and \code{cv_summary}, see details for its description. All arguments are optional, however \code{n_steps}, \code{n_windows} and \code{cv_summary} should only be used when \code{cv_update = TRUE}.}

\item{base_dates}{TRUE or FALSE, indicating whether initial date in modeling should be kept for update, if possible.}

\item{force_request}{TRUE or FALSE, indicating whether estimation should continue when not all packs can be updated.}

\item{outlier_update}{TRUE or FALSE, indicating whether a new search for outliers should be conducted. Search for outliers will be done only when \code{cv_update} = TRUE.}

\item{breakdown}{TRUE, FALSE or vector with 'row_id' of model to run breakdown (max of 3 models will be updated). If \code{breakdown} is TRUE, will run breakdown for the first 3 arima models, if a vector with 'row_id' is provided, the first 3 arima models in this list will be updated.}
}
\value{
None. Will break if any argument is not properly defined.
}
\description{
This function will perform an initial check if all necessary
arguments to run the \code{run_update} were provided.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
