% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_string_treat.R
\name{vec_string_treat}
\alias{vec_string_treat}
\title{Treats vector of string to have a clean name}
\usage{
vec_string_treat(var_names)
}
\arguments{
\item{var_names}{vector of variable names to be made into clean names.}
}
\value{
A vector of clean names.
}
\description{
This function will receive a vector of variable names, make them
into clean names. Later it will check for duplicates and append number to make
sure all names are different.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
