% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Getting url for modeling}
\usage{
get_url(type)
}
\arguments{
\item{type}{if 'models' or 'update'}

\item{run_local}{if TRUE, it will be sent to Docker to run locally. Default: FALSE.}
}
\value{
Api url.
}
\description{
Getting url for FaaS modeling.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
